execagg: extract out some aggregate helpers into a separate package
This allows us to break dependency of `colexecagg` on `execinfra`.

Release note: None
