+++ pkg/sql/BUILD.bazel
@@ -326,6 +326,7 @@ go_library(
         "//pkg/sql/distsql",
         "//pkg/sql/enum",
         "//pkg/sql/execinfra",
+        "//pkg/sql/execinfra/execagg",
         "//pkg/sql/execinfra/execopnode",
         "//pkg/sql/execinfrapb",
         "//pkg/sql/execstats",
