+++ pkg/sql/execinfra/execagg/BUILD.bazel
@@ -1,0 +1,16 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "execagg",
+    srcs = ["base.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/execinfra/execagg",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/sql/execinfrapb",
+        "//pkg/sql/sem/builtins",
+        "//pkg/sql/sem/eval",
+        "//pkg/sql/sem/tree",
+        "//pkg/sql/types",
+        "@com_github_cockroachdb_errors//:errors",
+    ],
+)
