+++ pkg/sql/execinfra/execagg/base.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package execinfra
+package execagg
 
 import (
 	"strings"
